refactor(library/StringSubset): moves out of "/customTypes" directory
