+++ src/library/Base64CharacterEncodedByteSequence/index.ts
@@ -6,7 +6,7 @@ import type {
   StringParsable,
 } from '@/library/Parser/string';
 
-import StringSubset from '@/library/customTypes/StringSubset';
+import StringSubset from '@/library/StringSubset';
 
 import {
   droppingLastCharacter,
