+++ src/library/customTypes/NonTrivialString/index.ts
@@ -4,7 +4,7 @@ import type {
   StringParsable,
 } from '@/library/Parser/string';
 
-import StringSubset from '@/library/customTypes/StringSubset';
+import StringSubset from '@/library/StringSubset';
 
 import {
   concatenated,
